Unify OIDC client construction and split analytics-web-srv auth.rs
- Add micromegas_auth::oidc_client with ConfiguredCoreClient type alias and
  DiscoveredProvider (discover + build_client), the single canonical home for
  login-flow provider discovery and client construction.
- Split analytics-web-srv/src/auth.rs into an auth/ submodule: mod.rs
  re-exports the flat public API, config.rs (OidcClientConfig), state.rs
  (AuthState), cookies.rs, claims.rs, handlers.rs.
- AuthState.oidc_provider now caches a DiscoveredProvider; get_oidc_provider
  delegates to DiscoveredProvider::discover; build_oidc_client and the
  web-local OidcProviderInfo/ConfiguredCoreClient are removed.
- Drop the two dead `let _client = build_oidc_client(...)` bindings in
  auth_callback/auth_refresh; auth_login now calls provider.build_client().
- Behavior-preserving: public API paths and all existing tests unchanged.
